Stop using next/image for images.
The build throws warnings for LCP and bandwidth, but the latter is definitely no issue, and the LCP would depend on serving this from the actual NodeJS webapp instead of the pre-built static site.

So ignore the warning, and use the standard <img> tag that works without needing dynamic loaders.